Módosítások

+++ database/migrations/2015_10_28_114913_create_horses_table.php
@@ -18,7 +18,7 @@ class CreateHorsesTable extends Migration
             $table->string('colour');
             $table->string('sex');
             $table->integer('age');
-            $table->integer('in_competition');
+            $table->tinyInteger('in_competition');
             $table->timestamps();
         });
     }
