+++ database/migrations/2015_11_05_151711_create_competitions_table.php
@@ -21,7 +21,7 @@ class CreateCompetitionsTable extends Migration
             $table->string('host');
             $table->date('date');
             $table->string('category');
-            $table->integer('in_competition');
+            $table->tinyInteger('in_competition');
             $table->timestamps();
         });
     }
