+++ database/migrations/2015_11_20_133750_add_two_columns_to_competitiongroups.php
@@ -16,7 +16,7 @@ class AddTwoColumnsToCompetitiongroups extends Migration
         {
             $table->integer('bouts_per_match');
             $table->integer('fencing_bouts');
-            $table->integer('entry_closed');
+            $table->tinyInteger('entry_closed');
             $table->text('riding_time_limit');
         });
     }
